Disable -fpermissive and fix all errors

+++ game_shared/stringlib.cpp
@@ -456,7 +456,7 @@ char *Q_stristr( const char *string, const char *string2 )
 
 int Q_vsnprintf( char *buffer, size_t buffersize, const char *format, va_list args )
 {
-	size_t	result;
+	int	result;
 
     result = vsnprintf( buffer, buffersize, format, args );
